+++ changelog.rst
@@ -6,6 +6,7 @@ Change Log
 - Unreleased
 
   - Fix undo session files using wrong extension (``.nil`` instead of writing ``.el`` files).
+  - Fix ``print-length`` & ``print-level`` settings corrupting saved undo data.
 
 - Version 0.7 (2024-07-07)